+++ CHANGELOG.md
@@ -8,4 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`8db0412...main`][8db0412...main].
 
+### Added
+
+- Added `Command\NormalizeCommand` and a `json-normalize` binary, normalizing formatting, as well as properties when a JSON schema is specified ([#696]), by [@localheinz]
+
 [8db0412...main]: https://github.com/ergebnis/json-normalize/compare/8db0412...main
+
+[#696]: https://github.com/ergebnis/json-normalize/pull/696
+
+[@localheinz]: https://github.com/localheinz